+++ packages/utils/src/index.ts
@@ -1,4 +1,4 @@
 export * from './errors';
 export * from './assert';
 export * from './string';
-export * from './cryptos';
+export * from './crypto';
